+++ bld/wl/c/loadpe.c
@@ -599,7 +599,7 @@ static unsigned_32 WriteExportInfo( pe_object *object, unsigned_32 file_align, p
     unsigned_32         size;
     pe_export_directory dir;
     const char          *name;
-    unsigned            namelen;
+    size_t              namelen;
     entry_export        **sort;
     entry_export        *exp;
     unsigned            i;
@@ -1122,7 +1122,7 @@ void FiniPELoadFile( void )
                 PE64( h ).dll_flags |= PE_DLL_PERTHRD_TERM;
             }
         }
-    
+
         if( FmtData.u.pe.lnk_specd ) {
             PE64( h ).lnk_major = FmtData.u.pe.linkmajor;
             PE64( h ).lnk_minor = FmtData.u.pe.linkminor;
@@ -1142,9 +1142,9 @@ void FiniPELoadFile( void )
             LnkMsg( WRN+MSG_VALUE_INCORRECT, "s", "alignment" );
             FmtData.u.os2.segment_shift = DEFAULT_SEG_SHIFT;
         }
-    
+
         PE64( h ).file_align = file_align;
-    
+
         if( FmtData.u.pe.osv_specd ) {
             PE64( h ).os_major = FmtData.u.pe.osmajor;
             PE64( h ).os_minor = FmtData.u.pe.osminor;
@@ -1152,7 +1152,7 @@ void FiniPELoadFile( void )
             PE64( h ).os_major = PE_OS_MAJOR;
             PE64( h ).os_minor = PE_OS_MINOR + 0xb;      // KLUDGE!
         }
-    
+
         PE64( h ).user_major = FmtData.major;
         PE64( h ).user_minor = FmtData.minor;
         if( FmtData.u.pe.sub_specd ) {
@@ -1278,7 +1278,7 @@ void FiniPELoadFile( void )
                 PE32( h ).dll_flags |= PE_DLL_PERTHRD_TERM;
             }
         }
-    
+
         if( FmtData.u.pe.lnk_specd ) {
             PE32( h ).lnk_major = FmtData.u.pe.linkmajor;
             PE32( h ).lnk_minor = FmtData.u.pe.linkminor;
@@ -1297,10 +1297,10 @@ void FiniPELoadFile( void )
             LnkMsg( WRN+MSG_VALUE_INCORRECT, "s", "alignment" );
             FmtData.u.os2.segment_shift = DEFAULT_SEG_SHIFT;
         }
-    
+
         file_align = 1UL << FmtData.u.os2.segment_shift;
         PE32( h ).file_align = file_align;
-    
+
         if( FmtData.u.pe.osv_specd ) {
             PE32( h ).os_major = FmtData.u.pe.osmajor;
             PE32( h ).os_minor = FmtData.u.pe.osminor;
@@ -1308,7 +1308,7 @@ void FiniPELoadFile( void )
             PE32( h ).os_major = PE_OS_MAJOR;
             PE32( h ).os_minor = PE_OS_MINOR + 0xb;      // KLUDGE!
         }
-    
+
         PE32( h ).user_major = FmtData.major;
         PE32( h ).user_minor = FmtData.minor;
         if( FmtData.u.pe.sub_specd ) {
